Suppress blind exception flake8 errors
I think it is acceptable to catch a generic Exception in both
places where the error is being reported.

In the command.py, we could use the exception type
`warnings._OptionError`, but I don't think it's safe to rely this
private exception type as it might not remain stable in future.
It would be better to not call the private function altogether,
but I'm not sure how best to avoid it.

In executor/__init__.py, we are propagating the exception in the
except block. In this case, I consider the linter is reporting a
false-positive, as it is unable to handle this situation.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>