Merge the pytest configuration
The addition of the pytest.ini is eclipsing the pytest configuration
that was already present in the setup.cfg.

There can be only one.

This moves the configuration from the pytest.ini into the setup.cfg and
re-adjusts the warnings filter to get things passing on all of Ubuntu,
macOS, and Windows.