Break apart topological_order_packages function
This function is currently doing a few discrete things:
1. Decorating descriptors
2. Adding recursive dependencies
3. Ordering decorators
4. Ordering recursive dependencies on decorators

There are already dedicated functions for performing the first two
tasks, so this change introduces a new topological_order_decorators
function which includes only the novel parts of
topological_order_packages and makes the original function call into the
three functions that accomplish the same tasks.

After teasing the duplicated code apart, the actual topological ordering
portion of the algorithm could be simplified and no longer required the
temporary object for storing pre-decorated descriptors.

There isn't really any reason to deprecate the original function, but
the improved code reuse results in better test coverage and the discrete
ordering function could be useful when ordering descriptors which are
already decoratord.