Fix the bootstrap test in GitHub Actions
Installing the test dependencies like this also enables importing from
the package itself, which is contrary to the bootstrapping scenario
we're trying to test. Simplest solution is to specifically uninstall it
again, which should leave all of the dependencies in place.

Additionally, the test results during the bootstrap build were not
causing the action to fail, and were therefore being ignored.

Furthermore, it appears that the COLCON_ALL_SHELLS variable is not
sufficient to generate usable bash hooks on Windows, so I'm disabling
the bash test until we can come up with a good way to support it.

Lastly, while in the package root, it's possible to perform imports from
colcon_core as if it were already installed. The bootstrap scenario
doesn't necessitate running from the repository root, so the test
shouldn't do so either.