Demote uncommon pytest extension packages to suggestions
These extensions are not packaged for Ubuntu Bionic or Focal, and while
it's still possible to install colcon with these recommendations
present, the package is excluded from a normal 'apt upgrade' because of
their absence. A 'dist-upgrade' works as expected, but we still want
users to be able to upgrade normally.

Fixes: f02d3ba4e5414ee0d17e00287b2ac58e69520ee3