Drop unneeded nonlocal from pytest hook (#509)
The nonlocal declaration is only needed when a variable is assigned to
within a tighter scope. If the variable is only referenced but is not
defined within the scope, Python should look for it in the higher scope
without the declaration.

This change makes colcon-core's pytest hook compatible with Python 2.
This is necessary to invoke Python 2 pytest while colcon is in a user's
`PYTHONPATH`, where we're currently seeing a syntax error.