Explicitly specify Python installation directories
Since we're overriding the Python installation layout scheme in colcon,
we shouldn't assume that setuptools is going to do the same thing. On
Ubuntu Jammy, the system's setuptools package has a patch that seems to
make it act how we want it to, but installing setuptools from pip
without the system's patch causes these paths to change.

Rather than continue to attempt to predict setuptools' behavior, we
should just specify the spots where we want the files manually.