Stop using 'echo' in shell test
On Linux, 'echo' is an executable. On Windows, 'echo' is a shell
command. Even if we enable the shell while executing this test, the
command acts differently on Windows and it is difficult to get it to
print newlines as is required by this test.

In any case, we can just use the python executable to print the text
instead, which is obviously available during testing.