Return list of generated files from create_environment_scripts
The create_environment_scripts function writes package-specific files to
disk based on available shell extensions. This change causes that
function, and the functions it calls in ShellExtension instances, to
return a list of files which were actually written.

This can be useful in downstream processes to understand what files were
installed as part of the package, which could facilitate uninstalling
those packages properly.

This change includes a minor version bump to the ShellExtensionPoint.
Existing implementations of ShellExtensionPoint will not be broken by
this change, but the files they write will not be included in the list
until they update to the new extension point API.

Consumers of the ShellExtensionPoint API are all updated within this
change.