Revert "Deprecate colcon_core.task.python.test.has_test_dependency()"
It turns out that this mechanism is used to communicate the Python-type
dependencies from the setuptools options when processing ament_python
packages, which otherwise have ROS-type dependencies. So while we're
seeing `python3-pytest` in the test dependency list, the pure python
code here looks for the pure python dependency `pytest`.

Since this doesn't actually save us much time and is clearly still
needed, I'll take the same approach in the Python project code as was
taken in ament_python so that we can all use the same Python testing
task.

This reverts commit 300f84132d2036b13e62ec429e7c3671d905b459.