Resolve path arguments when specified on command line
This change enhances the behavior of command line arguments which
specify a path which might be relative to the current directory.

The new functionality captures the current working directory when the
parser is constructed, and captures a given relative path on the command
line by resolving it from that captured working directory. If an
argument is left to the default value, it is not resolved by this new
functionality and is left for resolution in the consuming code later.

The main use case for this capture-then-resolve behavior is to support
changing the working directory of the colcon process. If that happens,
the context of where a user intended a relative path to be based is
lost, so this change captures the directory early in the colcon startup
sequence.