Track dependency categories on recursive dependencies
When enumerating the recursive dependencies for a package descriptor,
track which dependency categories the dependency appears to be a part
of. This information can be used later to drop certain dependencies or
inform a developer of the context around a relationship.

Note that the category list is not merged when a dependency appears
multiple times in a package's tree, and the most shallow instance
prevails.