Drop Windows event loop hack in Python 3.8+
In Python versions prior to 3.8, it appears that attempting to close the
event loop after a Ctrl-C would reliably hang the process, which would
need to be killed.

I've been unable to reproduce the behavior in any newer Python versions,
so I think it's time to set the timeline for removing the hack entirely.

This should take care of the common ResourceWarning messages when using
newer Python versions.