Implement a custom distutils command to symlink data_files
The default implementation of install_data will always copy files into
the destination directory. When we use the 'develop' command, we
actually need to specifically tell setuptools to do something with the
data_files or they will be ignored. Instead of telling setuptools to use
install_data as-is, we can implement a custom version of install_data
that will try to symlink the files instead.