Remove setup.py link after successful _undo_develop
This function reverts a previous setuptools 'develop' operation, but
doesn't actually clean up the artifacts under the build tree that it
used to determine that 'develop' was previously used. The result is that
after performing a build with --symlink-install, all subsequent builds
of that package without --symlink-install will run the commands to
"undo" the develop operation even if there is nothing to undo.

This should provide a small performance improvement under those specific
circumstances.