Use log, build, and install folders from colcon workspace root
Currently, colcon expects to be run from the root folder of a workspace.
To allow running from a subfolder, but still using the top-level log, build, and install folders,
get_root_path() traverses up the folder hierarchy until it finds a folder with a .colcon_root marker file
or the typical log, build, and install folders indicating a colcon workspace root.