Add Debian 'Replaces: colcon' field
Upstream Debian has decided to package colcon, and they created a
package simply called 'colcon' which provides only the /usr/bin/colcon
executable and some weak dependencies as a sort of replacement for
colcon-common-extensions. This conflicts with our python3-colcon-core
package during unpacking.

We could add 'Conflicts: colcon', but I think this is a great use case
for 'Replaces:' because the only file provided by 'colcon' is
/usr/bin/colcon, which we obviously provide here. This way, folks can
still successfully install 'colcon' and get the weak dependencies even
if our version of python-colcon-core is installed, and our copy of
/usr/bin/colcon will take precedence over the one provided by 'colcon'.