Add colcon_core.logging.get_effective_console_level function
When colcon routes log messages to log files at a different level from
the console, it makes it a little more convoluted to determine what log
level is actually set.

When we're utilizing non-colcon libraries that also use python's logging
module, we'll typically want to "synchronize" colcon's configured log
level with the other library. This function can be used to determine
what level colcon's console logging is set to.