Treat distutils as a 3rd party module (#704)
In Python 3.12 and newer, distutils is no longer part of the standard
library. It appears that flake8 has been updated to regard it as such,
so we need to update our import order.

Until the updated flake8 is ubiquitous, we'll need to specifically
suppress the linting for those older versions.