Add framework for enumerating installed Python files
In order for colcon to consume wheels produced by PEP 517 build
backends, we'll need to explicitly uninstall existing Python packages
from the install space. This primarily stems from the differences in
metadata representation, where the previous setuptools-only Python build
mechanisms wrote `.egg-link` and `.egg-info` metadata, and wheels will
use `.dist-info` metadata.

There should be only one metadata representation in the install space
for a given package. The legacy setuptools build logic already has a
special case for dealing with a switch from `.egg-info` to `.egg-link`
and vice-versa. As we toss `.dist-info` into the mix, the need arises
for a more robust solution.

This new functionality is not currently used by colcon, but will be used
as part of the transition to standards-based Python build support.